data/requirements.txt
PyYAML==6.0
passlib==1.7.4
fastapi==0.96.0
uvicorn==0.22.0
gunicorn==20.1.0
bcrypt==4.0.1
